% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{metrics}
\alias{metrics}
\title{Metrics of a model from the confusion matrix}
\usage{
metrics(real, predicted)
}
\arguments{
\item{real}{Values of the response variable in test set.}

\item{predicted}{Predicted values of the test set response variable in the model.}
}
\value{
Confusion matrix: TP FN FP TN. Recall, Precision, Accuracy, F1 Score.
}
\description{
metrics() provides the classification values associated with a confusion matrix, and various performance metrics.
}
\details{
The nomenclature of the values associated with the confusion matrix returned by the function are TP: true positive, FN: false negative, FP: false positive and TN: true negative.
}
