% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gain.R
\name{model_gain}
\alias{model_gain}
\title{Gain curve}
\usage{
model_gain(y_test, y_test_prob)
}
\arguments{
\item{y_test}{Values of the response variable in test set.}

\item{y_test_prob}{Probabilities of the response variable in test set.}
}
\description{
model_gain() returns the gain curve, which is constructed from the percentage of the tested population and the cumulative proportion of successes. The graph allows evaluating the percentage of gain in the prediction of success in relation to the percentage of the tested population.
}
