% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ks.R
\name{model_ks}
\alias{model_ks}
\title{KS curve: Kolmogorov - Smirnov curve}
\usage{
model_ks(y_train, y_train_prob, y_test, y_test_prob, type = 3)
}
\arguments{
\item{y_train}{Values of the response variable in training set.}

\item{y_train_prob}{Probabilities of the response variable in training set.}

\item{y_test}{Values of the response variable in test set.}

\item{y_test_prob}{Probabilities of the response variable in test set.}

\item{type}{1 indicates that the KS curve is plotted using the cumulative proportions of real zeros and ones, while 2 indicates that the FPR and TPR are used. Value 3 generates both graphs (default value).}
}
\description{
model_ks() returns the Kolmogorov-Smirnov graph, which seeks to compare the cumulative distributions of successes versus failures. It is possible to choose between different graphs to study this, the first illustrates the cumulative distributions, while the second uses the true positive ratio (TPR) and the false positive ratio (FPR). The graphs allow the discrimination capacity of the model to be evaluated.
}
