% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr.R
\name{model_pr}
\alias{model_pr}
\title{PR curve: Precision - Recall curve}
\usage{
model_pr(y_train, y_train_prob, y_test, y_test_prob, type = 3)
}
\arguments{
\item{y_train}{Values of the response variable in training set.}

\item{y_train_prob}{Probabilities of the response variable in training set.}

\item{y_test}{Values of the response variable in test set.}

\item{y_test_prob}{Probabilities of the response variable in test set.}

\item{type}{1 indicates that the PR curve is plotted using Recall on the x-axis and Precision on the y-axis, while 2 indicates that both Recall and Precision are plotted using probabilities on the x-axis. Value 3 generates both graphs (default).}
}
\description{
model_pr() returns the Precision - Recall curve, which is based on the confusion matrix (Precision and Recall). The graph allows you to study cases of population imbalance.
}
