% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc.R
\name{model_roc}
\alias{model_roc}
\title{ROC curve: receiver operating characteristic curve}
\usage{
model_roc(y_train, y_train_prob, y_test, y_test_prob)
}
\arguments{
\item{y_train}{Values of the response variable in training set.}

\item{y_train_prob}{Probabilities of the response variable in training set.}

\item{y_test}{Values of the response variable in test set.}

\item{y_test_prob}{Probabilities of the response variable in test set.}
}
\description{
model_roc() returns the ROC curve by calculating the true positive rate (TPR) and false positive rate (FPR), as well as the area under the curve. The graph allows you to evaluate the discrimination capacity of the model.
}
