% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{mtr.glm}
\alias{mtr.glm}
\title{Métricas de desempeño}
\usage{
mtr.glm(
  y,
  x,
  link = "logit",
  n = 0.8,
  balance = NULL,
  ms = NULL,
  semilla = 9999,
  plots = F,
  ref = FALSE
)
}
\arguments{
\item{y:}{variable de estudio binaria.}

\item{x:}{matriz de variables explicativas.}

\item{link:}{link del modelo GLM, c("logit", "probit", "cloglog", ...), "logit" por defecto.}

\item{n:}{número entre 0 y 1 que indica el porcentaje de partición del conjunto de entrenamiento, por defecto es 0.8.}

\item{balance:}{indica si se balancea la muestra de entrenamiento c("over", "under"), por defecto es NULL (no se balancea).}

\item{ms:}{método de selección de variables c("forward", "backward"), por defecto es NULL (todas las variables).}

\item{semilla:}{semilla para los procesos de aleatorios de balanceo, por defecto es 9999.}

\item{plots:}{indica si los gráficos de desempeño se muestran en el mismo archivo, por defecto es FALSE.}

\item{ref:}{por defecto es FALSE, indica si la categoría de referencia de la variable respuesta se debe invertir.}
}
\value{
ROC, PR, KS, metricas, conjunto test, conjunto train, modelo test, modelo train.
}
\description{
Métricas de desempeño
}
\examples{
n <- 500
x1 <- 2*runif(n) - 1
x2 <- 2*runif(n) - 1
x <- cbind(1, x1, x2)
beta <- c(-1, -3.0, 5.0)
probs <- exp(beta \%*\% t(x))/(1+exp(beta \%*\% t(x)))
y <- as.factor(rbinom(n,1,probs) + 1)
modelo = mtr.glm(y,x, link = "logit", balance = "under", plots = F, ms = "forward")
modelo$M.test
}
