% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MtrGLM-package.R
\docType{package}
\name{MtrGLM-package}
\alias{MtrGLM}
\alias{MtrGLM-package}
\title{MtrGLM: Generalized linear model metrics}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Package for graphs and performance metrics of a generalized linear model.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Dfranzani/MtrGLM}
  \item Report bugs at \url{https://github.com/Dfranzani/MtrGLM/issues}
}

}
\author{
\strong{Maintainer}: Daniel Franzani \email{dfranzani@gmail.com} (\href{https://orcid.org/0009-0006-7384-1195}{ORCID})

}
\keyword{internal}
