% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sleep_data}
\alias{sleep_data}
\title{Sleep Health and Lifestyle Dataset}
\format{
A data frame with 374 rows and 9 variables:
\describe{
\item{\code{Gender}}{Gender of the person (Male/Female).}
\item{\code{Age}}{Person's age in years.}
\item{\code{Sleep.Duration}}{Number of hours the person sleeps per day.}
\item{\code{Quality.of.Sleep}}{A subjective rating of sleep quality, which takes a discrete value from 1 to 10.}
\item{\code{Physical.Activity.Level}}{Number of minutes/day that the person performs physical activity daily.}
\item{\code{Stress.Level}}{A subjective rating of the level of stress experienced by the person, which takes a discrete value ranging from 1 to 10. A higher value indicates a higher level of stress.}
\item{\code{Heart.Rate}}{The person's resting heart rate in beats per minute.}
\item{\code{Daily.Steps}}{Number of steps the person takes per day.}
\item{\code{Sleep.Disorder}}{Presence or absence of a sleep disorder in the person (None, Insomnia, Sleep Apnea).}
}
}
\source{
Original database: \url{https://www.kaggle.com/datasets/uom190346a/sleep-health-and-lifestyle-dataset}

Modified data subset: \url{https://raw.githubusercontent.com/Dfranzani/Bases-de-datos-para-cursos/main/2023-1/sleep2.csv}
}
\usage{
sleep_data
}
\description{
The database is an extract of another database extracted from a publication on Kaggle (public domain). The sleep health and lifestyle dataset consists of 374 rows and 8 columns, covering a wide range of variables related to sleep and daily habits. It includes details such as gender, age, occupation, sleep duration, sleep quality, physical activity level, stress levels, heart rate, daily steps, and the presence or absence of sleep disorders.
}
\keyword{datasets}
